+++ CHANGELOG.md
@@ -5,6 +5,16 @@ Objet de la version ``2.0.0`` : Concevoir un coeur du moteur, une interface pour
 L'objectif est aussi de transformer le moteur en respectant les **standards PSR**.
 
 
+## 2.0.x-alpha3 (2017-10-07)
+
+- [CHANGED] ``addLanguage`` devient ``addLanguages``, car admet un ou plusieurs code de langue.
+- [CHANGED] ``removeLanguage`` devient `removeLanguages`,car admet un ou plusieurs code de langue.
+- [CHANGED] ``compile`` devient `deploy` qui correspond plus à l'aspect fonctionnel.
+- [ADDED] Ajout de la méthode ``checkCode`` qui valide le format réprésentant un code de langue.
+- [ADDED] Ajout de la méthode ``setRefLanguage`` qui définie la langue de référence.
+- [ADDED] Ajout de la méthode ``getRefLanguage`` qui retourne le pack de langue de référence.
+
+
 ## 2.0.x-alpha2 (2017-10-07)
 
 - [CHANGED] ``Compilator`` devient ``Compiler``.
@@ -14,7 +24,8 @@ L'objectif est aussi de transformer le moteur en respectant les **standards PSR*
 
 - [ADDED] Création d'un espace de nom : ``SYSLang``.
 - [CHANGED] ``SYSLangCompilator`` devient ``Compilator`` et fait office de coeur.
-- [CHANGED] ``SYSLang`` reste ``SYSLang``, mais n'a de rôle que pour l'utilisation sur serveurs Web sous PHP et n'est plus le coeur.
+- [CHANGED] ``SYSLang`` reste ``SYSLang``, mais n'a de rôle que pour l'utilisation sur serveurs Web sous PHP 
+et n'est plus le coeur.
 - [CHANGED] ``get_avail_languages`` devient ``getRegLanguages`` dans la classe ``Compilator``.
 - [CHANGED] ``list_languages`` devient ``listLanguages`` dans la classe ``Compilator``.
 - [CHANGED] ``save_xml`` devient ``saveXml`` dans la classe ``Compilator``.